+++ antarest/study/storage/variantstudy/model/command/remove_user_resource.py
